refactor: extract system file bootstrap into sub-module, update readability baseline
Moves find_system_files() and bootstrap_system_files() into
os_apps::system_files to keep mod.rs under the readability ratchet.
Unwrap count unchanged (ci-ok unwrap replaced with expect()).
