+++ .ci/readability-baseline.env
@@ -1,9 +1,9 @@
 # Generated by scripts/readability-ratchet.sh
-PROD_RS_TOTAL=356
+PROD_RS_TOTAL=357
 PROD_FILES_GT300=115
 PROD_FILES_GT500=48
 PROD_FILES_GT1000=2
-PROD_MAX_FILE_LINES=2875
+PROD_MAX_FILE_LINES=2889
 PROD_MAX_FILE_PATH=crates/temper-platform/src/os_apps/mod.rs
 ALLOW_CLIPPY_COUNT=23
 ALLOW_DEAD_CODE_COUNT=11
